Add CreateSafeContext util and use it within channels provider, refactor
