--- components/ChannelsProvider/useChannels.ts
@@ -1,11 +1,0 @@
-import { useContext } from "react";
-
-import { ChannelsConsumerProps, ChannelsContext } from "./ChannelsProvider";
-
-const useChannels = (): ChannelsConsumerProps => {
-  const channelsConsumerProps = useContext(ChannelsContext);
-
-  return channelsConsumerProps;
-};
-
-export default useChannels;
