+++ components/SaleorProviderWithChannels/SaleorProviderWithChannels.tsx
@@ -2,8 +2,7 @@ import { SaleorProvider } from "@saleor/sdk";
 import React, { PropsWithChildren, useEffect } from "react";
 
 import { saleorClient } from "@/lib/graphql";
-
-import useChannels from "../ChannelsProvider/useChannels";
+import { useChannels } from "../ChannelsProvider";
 
 const SaleorProviderWithChannels = ({ children }: PropsWithChildren<{}>) => {
   const { currentChannel } = useChannels();
