+++ lib/useSafeContext.ts
@@ -1,0 +1,17 @@
+import React from "react";
+
+function createSafeContext<TValue extends {} | null>() {
+  const context = React.createContext<TValue | undefined>(undefined);
+
+  function useContext() {
+    const value = React.useContext(context);
+    if (value === undefined) {
+      throw new Error("useContext must be inside a Provider with a value");
+    }
+    return value;
+  }
+
+  return [useContext, context.Provider] as const;
+}
+
+export default createSafeContext;
